requirements.txt
pandas
numpy
inflect
pyspellchecker
contractions
emot
flashtext
fasttext
hydra-core
torch
tensorboard
transformers